+++ src/__shared__/ui/shadcn/dialog.tsx
@@ -1,10 +1,9 @@
 "use client";
 
+import { cn } from "@/__shared__/utils/cn";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { X } from "lucide-react";
 import * as React from "react";
-
-import { cn } from "@/__shared__/utils/utils";
 
 const Dialog = DialogPrimitive.Root;
