+++ src/__shared__/ui/shadcn/label.tsx
@@ -1,10 +1,9 @@
 "use client";
 
+import { cn } from "@/__shared__/utils/cn";
 import * as LabelPrimitive from "@radix-ui/react-label";
 import { cva, type VariantProps } from "class-variance-authority";
 import * as React from "react";
-
-import { cn } from "@/__shared__/utils/utils";
 
 const labelVariants = cva(
   "text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
