+++ src/__shared__/ui/shadcn/radio-group.tsx
@@ -1,10 +1,9 @@
 "use client";
 
+import { cn } from "@/__shared__/utils/cn";
 import * as RadioGroupPrimitive from "@radix-ui/react-radio-group";
 import { Circle } from "lucide-react";
 import * as React from "react";
-
-import { cn } from "@/__shared__/utils/utils";
 
 const RadioGroup = React.forwardRef<
   React.ElementRef<typeof RadioGroupPrimitive.Root>,
