+++ src/__shared__/ui/shadcn/scroll-area.tsx
@@ -3,7 +3,7 @@
 import * as ScrollAreaPrimitive from "@radix-ui/react-scroll-area";
 import * as React from "react";
 
-import { cn } from "@/__shared__/utils/utils";
+import { cn } from "@/__shared__/utils/cn";
 
 const ScrollArea = React.forwardRef<
   React.ElementRef<typeof ScrollAreaPrimitive.Root>,
