+++ src/__shared__/ui/shadcn/toast.tsx
@@ -1,9 +1,8 @@
+import { cn } from "@/__shared__/utils/cn";
 import * as ToastPrimitives from "@radix-ui/react-toast";
 import { cva, type VariantProps } from "class-variance-authority";
 import { X } from "lucide-react";
 import * as React from "react";
-
-import { cn } from "@/__shared__/utils/utils";
 
 const ToastProvider = ToastPrimitives.Provider;
